Add DocumentAwareMixin and refactor kernel clients
The kernel monitor extension v0.3.0 requires kernel clients to have an
is_beating() method. This was causing AttributeErrors with DocumentAwareKernelClient.

Solution:
- Created DocumentAwareMixin with all document-aware functionality (YRoom integration,
  output processing, cell state tracking)
- Refactored DocumentAwareKernelClient to use the mixin (reduced from 220 to 33 lines)
- Mixin can be reused by any kernel client (ZMQ or Gateway)
- All imports moved to top of files (no inline imports)

The mixin provides:
- YRoom management for collaborative editing
- Output processor with proper parent chain for session lookups
- Cell execution state tracking
- Document message handling (kernel_info, status, execute_input, outputs)
