+++ jupyter_server_documents/__init__.py
@@ -11,6 +11,7 @@ except ImportError:
 
 from .app import ServerDocsApp
 from .events import JSD_AWARENESS_EVENT_URI, JSD_ROOM_EVENT_URI
+from .document_aware_mixin import DocumentAwareMixin
 
 
 def _jupyter_labextension_paths():
